requirements.txt
absl-py==1.3.0
aiohttp==3.8.3
aiosignal==1.2.0
anyio==3.6.2
argo-workflows==3.6.1
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
astunparse==1.6.3
async-timeout==4.0.2
attrdict==2.0.1
attrs==22.1.0
Babel==2.11.0
backcall==0.2.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.6.3
bleach==5.0.1
bokeh==3.0.1
boto3==1.17.11
botocore==1.20.112
build==0.9.0
cachetools==4.2.4
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
commonmark==0.9.1
contourpy==1.0.6
cycler==0.11.0
daiquiri==3.2.1
decorator==5.1.1
defusedxml==0.7.1
delegator.py==0.1.1
dictdiffer==0.9.0
distlib==0.3.6
distro==1.8.0
entrypoints==0.4
fastjsonschema==2.16.2
filelock==3.8.0
Flask==2.2.2
flatbuffers==2.0.7
frozenlist==1.3.1
gast==0.4.0
gitdb==4.0.9
GitPython==3.1.29
google-auth==1.35.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.50.0
h5py==3.7.0
idna==3.4
importlib-metadata==5.0.0
importlib-resources==5.10.0
invectio==0.2.2
ipykernel==5.5.5
ipython==7.16.1
ipython-genutils==0.2.0
ipywidgets==8.0.2
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.2
jmespath==0.10.0
joblib==1.2.0
json5==0.9.10
jsonformatter==0.3.1
jsonschema==4.17.0
jupyter-bokeh==3.0.4
jupyter-lsp==1.5.1
jupyter-resource-usage==0.6.1
jupyter-server==1.4.1
jupyter-server-mathjax==0.2.6
jupyter-server-proxy==3.2.1
jupyter_client==7.4.4
jupyter_core==4.11.2
jupyterlab==3.2.4
jupyterlab-git==0.30.1
jupyterlab-lsp==3.10.0
jupyterlab-pygments==0.2.2
jupyterlab-s3-browser==0.10.1
jupyterlab-server==2.3.0
jupyterlab-widgets==3.0.3
kafka-python==2.0.2
keras==2.7.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
kubernetes==11.0.0
libclang==14.0.6
lxml==4.9.1
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.4.1
micropipenv==1.4.4
mistune==2.0.4
mock==4.0.3
multidict==6.0.2
nbclassic==0.2.6
nbclient==0.7.0
nbconvert==7.2.3
nbdime==3.1.0
nbformat==5.7.0
nest-asyncio==1.5.6
notebook==6.4.1
numpy==1.19.2
oauthlib==3.2.2
openshift==0.11.0
opt-einsum==3.3.0
packaging==21.3
pandas==1.2.4
pandocfilters==1.5.0
parso==0.8.3
pep517==0.13.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.3.0
pip-tools==6.8.0
pkgutil_resolve_name==1.3.10
platformdirs==2.5.2
plotly==5.5.0
prometheus-client==0.15.0
prompt-toolkit==3.0.32
protobuf==3.19.6
psutil==5.9.3
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyelftools==0.29
Pygments==2.13.0
pyparsing==3.0.9
pyrsistent==0.19.2
python-dateutil==2.8.2
python-json-logger==2.0.4
python-string-utils==1.0.0
pytz==2022.6
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
pyzmq==24.0.1
requests==2.28.1
requests-oauthlib==1.3.1
rfc5424-logging-handler==1.4.3
rich==12.6.0
rich-click==1.5.2
rsa==4.9
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
s3transfer==0.3.7
scikit-learn==0.24.1
scipy==1.6.2
semantic-version==2.10.0
Send2Trash==1.8.0
sentry-sdk==1.10.1
simpervisor==0.4
singleton-decorator==1.0.0
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
tenacity==8.1.0
tensorboard==2.6.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow-estimator==2.7.0
tensorflow-gpu==2.7.4
tensorflow-io-gcs-filesystem==0.27.0
termcolor==2.1.0
termcolor-whl==1.1.2
terminado==0.17.0
thamos==1.28.0
thoth-analyzer==0.1.8
thoth-common==0.36.6
thoth-python==0.16.11
threadpoolctl==3.1.0
tinycss2==1.2.1
toml==0.10.2
tomli==2.0.1
tornado==6.2
traitlets==5.5.0
typing_extensions==4.4.0
tzdata==2022.6
tzlocal==4.2
urllib3==1.26.12
virtualenv==20.16.6
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.4.2
Werkzeug==2.2.2
widgetsnbextension==4.0.3
wrapt==1.14.1
xyzservices==2022.9.0
yarl==1.8.1
yaspin==2.2.0
zipp==3.10.0
